backend/requirements.txt
fastapi>=0.104.0
pydantic>=2.5.0
uvicorn[standard]>=0.24.0
google-adk>=1.0.0
python-dotenv>=1.0.0
uuid>=1.30
crewai-tools<=0.1.16
beautifulsoup4